Deduce submodule update versions automatically
The nested changelog support currently needs to be provided with the
version range on the commit body, similarly to:

  Update <submodule> from <initialVersion> to <targetVersion>

This is cumbersome and does not work well with tools that automatically
update submodule dependencies like renovatebot.

This commit keeps the legacy behaviour but will extract the version range
from git metadata when not provided. This new behaviour can be used by
specifyng a commit body similar to:

  Update <submodule>

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>